requirements.txt
keyboard~=0.13.5
Pillow~=9.2.0
PyAutoGUI~=0.9.53
